Day 11: Capstone Project: Blackjack
A simplified Blackjack game, consolidating everything learnt up until this point. Very happy that I managed to complete this without leaning on any of the hints that Angela provides... I'm not sure if it's the most efficient code written, but as far as I can tell it works and I'm pleased with the outcome.