Day 14: Higher or Lower
As I get more confident with Python, the next steps for me will be to create code that not only functions, but is efficient. My code works (I think!) but it is a lot longer than it needs to be. I believe getting used to return statements and function inputs will help me with this.